Ticket #956: client wrapper: minor fixes after MR peer review

+++ tests/ClientServer/validation_tests/client_wrapper/check_wrapper_reverse_connection.c
@@ -68,7 +68,7 @@ START_TEST(test_wrapper_reverse_connections)
         .endpointUrl = INVALID_URL,
         .serverUri = NULL,
         .isReverseConnection = true,
-        .reverseConnectionConfigId = reverseEpId,
+        .reverseConnectionConfigId = (uint32_t) reverseEpId,
     };
 
     /* invalid server URL */
